Moved user-notifications
