+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/navigation-right/buttons.tsx
@@ -14,7 +14,7 @@
  **/
 import * as React from 'react'
 import { Button, buttonTypeEnum } from '../presentation/button'
-import UserNotifications from '../presentation/user-notifications/user-notifications'
+import UserNotifications from '../user-notifications'
 import UserView from '../user'
 
 const HelpView = require('../../component/help/help.view.js')
