+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/user-notifications/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './user-notifications'
